Implemented parsing of shell commands and parsing over multiple lines.
Divided antlr folder into multiple grammar rules so that WACCShellParser.g4 grammar does not have the prog node (since prog node won't be used in the shell) and instead has command which is the starting node for shell commands.

ASTGeneratorShellVisitor.kt is needed for the new shell grammar which is mostly duplicate with ASTGeneratorVisitor.kt and should be changed later to get rid of all the duplicate code.

Added shell class and the runInteractiveShell method to work such that it can also read input from a different source such as a file.

Added testing of parsing all the valid code examples using the shell which all pass.